Change return type of ft_pow from int to size_t

+++ libft.h
@@ -62,7 +62,7 @@ int		ft_atoi(const char *str);
 int		ft_abs(int nbr);
 char	*ft_itoa(int n);
 void	ft_intswap(int *x, int *y);
-int		ft_pow(int base, int power);
+size_t	ft_pow(int base, int power);
 
 //Lists
